Move TemporarySynchronizationContext to TestUtils.cs
